[frontend] refactor: use combobox for add filter component
